+++ go.mod
@@ -1,3 +1,15 @@
 module github.com/strick-j/cybr_pam_scim
 
 go 1.18
+
+require (
+	golang.org/x/exp v0.0.0-20220426173459-3bcf042a4bf5
+	golang.org/x/oauth2 v0.0.0-20220411215720-9780585627b5
+)
+
+require (
+	github.com/golang/protobuf v1.4.2 // indirect
+	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd // indirect
+	google.golang.org/appengine v1.6.6 // indirect
+	google.golang.org/protobuf v1.25.0 // indirect
+)
