+++ pkg/cybr_pam_scim/types/container_permissions.go
