+++ pkg/cybr_pam_scim/types/privileged_data.go
