+++ pkg/cybr_pam_scim/types/shared.go
